module6/svc/pyproject.toml
[project]
name = "configuration-service"
version = "0.1.0"
description = "Centralized configuration management service"
requires-python = ">=3.13.5"
dependencies = [
    "fastapi==0.116.1",
    "pydantic==2.11.7",
    "pydantic-settings>=2.0.0,<3.0.0",
    "pydantic-extra-types>=2.0.0,<3.0.0",
    "python-ulid>=2.0.0,<3.0.0",
    "psycopg2-binary==2.9.10",
    "uvicorn>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-cov==6.2.1",
    "pytest-asyncio>=0.20.0",
    "httpx==0.28.1",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
]
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-psycopg2>=0.41b0",
    "prometheus-client>=0.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["svc"]

[tool.pytest.ini_options]
testpaths = ["svc"]
python_files = ["*_test.py"]
addopts = "--cov=svc --cov-report=term-missing --cov-fail-under=80"

[dependency-groups]
dev = [
    "httpx==0.28.1",
    "pytest==8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov==6.2.1",
]
